tools/requirements.txt
pycryptodome==3.4.6
scrypt==0.8.0
